back/Adopt_Me/requirements.txt
appnope==0.1.0
asgiref==3.6.0
backcall==0.1.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
distlib==0.3.6
Django==3.2.18
djangorestframework==3.14.0
filelock==3.11.0
idna==2.10
importlib-metadata==6.3.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
platformdirs==3.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pygame==2.3.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.24.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.4.3
traitlets==4.3.2
typing_extensions==4.5.0
urllib3==1.25.9
virtualenv==20.21.0
wcwidth==0.1.7
zipp==3.15.0
